Improved DMA buffers + Improved ESI Prodigy 7.1 HiFi and ESI MAYA44 support + code refactoring. - Added a 64 bit target with a dedicated info.plist file to ensure compatibility with modern macOS versions. - Code refactoring to improve compatibility with 64 bit targets. - Replaced the DMA buffer allocation and management code to ensure that dma buffers are allocated in the first 2gb of ram to ensure compatibility with old pci chipsets. - Improved support for the ESI/Audiotrak Prodigy 7.1 HiFi. - Improved support for the ESI MAYA44 PCI.

+++ audiophile192.cpp
@@ -24,7 +24,7 @@
 static void
 GPIOWrite(struct CardData *card, unsigned long pos, unsigned long bit)
 {
-  unsigned long data = GetGPIOData(card->pci_dev, card->iobase);
+  UInt32 data = GetGPIOData(card->pci_dev, card->iobase);
 
   bit = (bit != 0);
 
@@ -371,8 +371,7 @@ ap192_set_rate (struct CardData *card, unsigned long speed)
 }
 
 static void
-ap192_audio_ioctl (struct CardData *card, int cmd,
-		   int *arg)
+ap192_audio_ioctl (struct CardData *card, int cmd, int *arg)
 {
 #if 0
   int left, right, value;
